requirements-dev.txt
# Fix: Q-110
# Development-time tooling dependencies.
pytest-cov>=4.1.0

# Fix: testing_ci_security_scanners
bandit[toml]>=1.7.8
pip-audit>=2.7.2
# TODO(codex): gitleaks must be installed separately because no PyPI package exists.
